requirements.txt
Django==3.0.4
django-crispy-forms==1.9.0
Pillow==7.1.1
six==1.14.0
selenium==3.141.0